Adds support for user-specifiable shaders via the "Set Shader" patch. There is a protocol now for displaying settings for a patch. Tap on it and it should appear alongside "Delete"
This is a very special changeset for me. Woolly Mammoth can now be used for non-trivial experimenting and programming.

At this very early stage, you need to make a "Render Output" node to be able to render anything out. Should be found under "Other." In the future this will be an undeleteable node with a preview of the output and some settings.